commotalkie-12: Create a driver for EByte * Driver receive returning value has to be an int

+++ include/driver/Driver.h
@@ -66,7 +66,7 @@ Driver Driver_Create(PinMap pins, RadioParams *params, IOCallback *io,
 unsigned long Driver_Send(Driver *driver, const Destination *destination,
                           const char *content, unsigned long size);
 
-long Driver_Receive(Driver *driver, char *buffer, unsigned long size);
+int Driver_Receive(Driver *driver, char *buffer, unsigned long size);
 
 void Driver_TurnOn(Driver *driver);
 void Driver_TurnOff(Driver *driver);
